+++ src/bin/run_mavis.py
@@ -22,7 +22,7 @@ def get_parser():
     parser.add_argument('-l', '--log-path', help='Output file for log messages; defaults to STDERR')
     parser.add_argument('-q', '--quiet', action='store_true', help='Logging for error messages only')
     parser.add_argument('-r', '--ready', action='store_true', help='Ready the working directory; find, link and filter inputs, and write Crowmell config. Cannot combine with --execute.')
-    parser.add_argument('-s', '--study', metavar='ID', help='Study ID, eg. PASS01')
+    parser.add_argument('-S', '--study', metavar='ID', help='Study ID, eg. PASS01')
     parser.add_argument('-v', '--verbose', action='store_true', help='Verbose logging')
     parser.add_argument('-w', '--work-dir', metavar='DIR', required=True, help='Working directory for input/output files')
     parser.add_argument('-x', '--execute', action='store_true', help='Launch Mavis job on Cromwell server. Cannot combine with --ready.')