+++ src/bin/sequenza_explorer.py
@@ -20,23 +20,23 @@ READ = 'read'
 def locate_sequenza_path(args):
     path_validator().validate_input_file(args.file_provenance)
     log_level = logger.get_args_log_level(args)
-    reader = provenance_reader(args.file_provenance, args.project, args.donor, log_level)
+    reader = provenance_reader(args.file_provenance, args.study, args.donor, log_level)
     return reader.parse_sequenza_path()
 
 def get_parser():
     parser = argparse.ArgumentParser(
         description='sequenza_explorer: Find Sequenza results, review purity/ploidy and apply the CGI heuristic to choose a gamma parameter'
     )
-    parser.add_argument('--debug', action='store_true', help='More verbose logging')
-    parser.add_argument('--verbose', action='store_true', help='Verbose logging')
-    parser.add_argument('--quiet', action='store_true', help='Logging for error messages only')
-    parser.add_argument('--log-path', help='Output file for log messages; defaults to STDERR')
+    parser.add_argument('-d', '--debug', action='store_true', help='More verbose logging')
+    parser.add_argument('-v', '--verbose', action='store_true', help='Verbose logging')
+    parser.add_argument('-q', '--quiet', action='store_true', help='Logging for error messages only')
+    parser.add_argument('-l', '--log-path', help='Output file for log messages; defaults to STDERR')
     subparsers = parser.add_subparsers(title='subcommands', help='sub-command help', dest='subparser_name')
     locate_parser = subparsers.add_parser(LOCATE, help='Locate Sequenza results in file provenance')
-    locate_parser.add_argument('-d', '--donor', metavar='DONOR', required=True, help='Donor ID for file provenance search')
+    locate_parser.add_argument('-D', '--donor', metavar='DONOR', required=True, help='Donor ID for file provenance search')
     locate_parser.add_argument('-f', '--file-provenance', metavar='PATH', default='/.mounts/labs/seqprodbio/private/backups/seqware_files_report_latest.tsv.gz',
                                help='Path to file provenance report; defaults to latest production report')
-    locate_parser.add_argument('-p', '--project', metavar='PROJECT', required=True, help='Project name for file provenance search')
+    locate_parser.add_argument('-S', '--study', metavar='STUDY', required=True, help='Study name for file provenance search')
     read_parser = subparsers.add_parser(READ, help='Read and summarize a Sequenza result .ZIP file')
     read_parser.add_argument('-i', '--in', metavar='PATH', required=True, dest='in_path', help='Path to .ZIP archive of Sequenza results')
     read_parser.add_argument('-j', '--json', metavar='PATH', help='Path for JSON output of gamma parameters and purity/ploidy')