+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # CHANGELOG
 
-## Unreleased
+## v0.2.4: 2022-03-23
 
 ### Fixed
 - GCGI-331 Standardize on --study for script options